changes to assist with debugging do_all_image_rna.sh - SEEMS OK

+++ do_all.sh
@@ -335,12 +335,12 @@ if [[ ${SKIP_TILING} == "False" ]];
         #~ echo "DO_ALL.SH: INFO: 'image' mode, so deleting saved image indices:  train_inds_image, test_inds_image"
         rm ${DATA_DIR}/train_inds_image  > /dev/null 2>&1
         rm ${DATA_DIR}/test_inds_image   > /dev/null 2>&1
-        #~ echo "DO_ALL.SH: INFO: recursively deleting files (tiles)           matching this pattern:  '*.png'                            <<< for image mode, deleting all the .png files (i.e. tiles) can take quite some time as there can be up to millions of tiles"
+        #~ echo "DO_ALL.SH: INFO: recursively deleting files (tiles)           matching this pattern:  '*.png'               <<< for image mode, deleting all the .png files (i.e. tiles) can take quite some time as there can be up to millions of tiles"
         find ${DATA_DIR} -type f -name *.png                                            -delete
     fi
     
-    if [[ ${INPUT_MODE} == "rna" ]]; then
-        #~ echo "DO_ALL.SH: INFO: 'image' mode, so deleting saved image indices:  train_inds_image, test_inds_image"
+    if [[ ${INPUT_MODE} == "rna" ]]  &&  [[ ${MULTIMODE} != 'image_rna' ]];
+        #~ echo "DO_ALL.SH: INFO: 'image' mode, so deleting saved rna indices:  train_inds_rna, test_inds_rna"
         rm ${DATA_DIR}/train_inds_rna    > /dev/null 2>&1
         rm ${DATA_DIR}/test_inds_rna     > /dev/null 2>&1
